refactors folium_viz_system folder structure
