+++ mescal/visualizations/folium_viz_system/base_system.py
@@ -7,7 +7,7 @@ import pandas as pd
 
 from mescal.kpis import KPICollection, KPI
 from mescal.typevars import ResolvedStyleType, StyleResolverType
-from mescal.visualizations.folium_viz_system.folium_generators import TooltipGenerator, PopupGenerator
+from mescal.visualizations.folium_viz_system.simple_generators import TooltipGenerator, PopupGenerator
 from mescal.visualizations.folium_viz_system.map_data_item import MapDataItem, ModelDataItem, KPIDataItem
