+++ mescal/visualizations/folium_viz_system/system_circle_marker.py
@@ -4,7 +4,7 @@ from typing import Type
 import folium
 
 from mescal.visualizations.folium_viz_system.map_data_item import MapDataItem
-from mescal.visualizations.folium_viz_system.system_base import ResolvedStyle, StyleResolver, StyleMapper, \
+from mescal.visualizations.folium_viz_system.base_system import ResolvedStyle, StyleResolver, StyleMapper, \
     FoliumObjectGenerator
