+++ mescal/visualizations/folium_viz_system/system_lines.py
@@ -6,10 +6,10 @@ import folium
 from folium.plugins import AntPath, PolyLineOffset
 from shapely import LineString
 
-from mescal.visualizations.folium_viz_system.folium_generators import TooltipGenerator, PopupGenerator, logger
+from mescal.visualizations.folium_viz_system.simple_generators import TooltipGenerator, PopupGenerator, logger
 from mescal.visualizations.folium_viz_system.map_data_item import MapDataItem
 
-from mescal.visualizations.folium_viz_system.system_base import ResolvedStyle, StyleResolver, StyleMapper, \
+from mescal.visualizations.folium_viz_system.base_system import ResolvedStyle, StyleResolver, StyleMapper, \
     FoliumObjectGenerator
